Move CellAttachedData and DataTransfer to Triangulation base class
