+++ include/deal.II/distributed/cell_weights.h
@@ -159,7 +159,7 @@ namespace parallel
      */
     static std::function<unsigned int(
       const typename dealii::Triangulation<dim, spacedim>::cell_iterator &cell,
-      const typename dealii::Triangulation<dim, spacedim>::CellStatus status)>
+      const CellStatus status)>
     make_weighting_callback(const DoFHandler<dim, spacedim> &dof_handler,
                             const WeightingFunction &weighting_function);
 
@@ -219,7 +219,7 @@ namespace parallel
     static unsigned int
     weighting_callback(
       const typename dealii::Triangulation<dim, spacedim>::cell_iterator &cell,
-      const typename dealii::Triangulation<dim, spacedim>::CellStatus status,
+      const CellStatus                                  status,
       const DoFHandler<dim, spacedim> &                 dof_handler,
       const parallel::TriangulationBase<dim, spacedim> &triangulation,
       const WeightingFunction &                         weighting_function);
