+++ include/deal.II/distributed/solution_transfer.h
@@ -349,7 +349,7 @@ namespace parallel
       std::vector<char>
       pack_callback(
         const typename Triangulation<dim, spacedim>::cell_iterator &cell,
-        const typename Triangulation<dim, spacedim>::CellStatus     status);
+        const CellStatus                                            status);
 
       /**
        * A callback function used to unpack the data on the current mesh that
@@ -359,7 +359,7 @@ namespace parallel
       void
       unpack_callback(
         const typename Triangulation<dim, spacedim>::cell_iterator &cell,
-        const typename Triangulation<dim, spacedim>::CellStatus     status,
+        const CellStatus                                            status,
         const boost::iterator_range<std::vector<char>::const_iterator>
           &                        data_range,
         std::vector<VectorType *> &all_out,
