+++ include/deal.II/particles/particle_handler.h
@@ -1143,7 +1143,7 @@ namespace Particles
     std::vector<char>
     pack_callback(
       const typename Triangulation<dim, spacedim>::cell_iterator &cell,
-      const typename Triangulation<dim, spacedim>::CellStatus     status) const;
+      const CellStatus                                            status) const;
 
     /**
      * Called by listener functions after a refinement step for each cell
@@ -1152,7 +1152,7 @@ namespace Particles
     void
     unpack_callback(
       const typename Triangulation<dim, spacedim>::cell_iterator &cell,
-      const typename Triangulation<dim, spacedim>::CellStatus     status,
+      const CellStatus                                            status,
       const boost::iterator_range<std::vector<char>::const_iterator>
         &data_range);
