+++ source/distributed/cell_weights.cc
@@ -128,7 +128,7 @@ namespace parallel
   template <int dim, int spacedim>
   std::function<unsigned int(
     const typename dealii::Triangulation<dim, spacedim>::cell_iterator &cell,
-    const typename dealii::Triangulation<dim, spacedim>::CellStatus     status)>
+    const CellStatus                                                    status)>
   CellWeights<dim, spacedim>::make_weighting_callback(
     const DoFHandler<dim, spacedim> &dof_handler,
     const typename CellWeights<dim, spacedim>::WeightingFunction
@@ -143,19 +143,17 @@ namespace parallel
       ExcMessage(
         "parallel::CellWeights requires a parallel::TriangulationBase object."));
 
-    return
-      [&dof_handler, tria, weighting_function](
-        const typename dealii::Triangulation<dim, spacedim>::cell_iterator
-          &                                                             cell,
-        const typename dealii::Triangulation<dim, spacedim>::CellStatus status)
-        -> unsigned int {
-        return CellWeights<dim, spacedim>::weighting_callback(
-          cell,
-          status,
-          std::cref(dof_handler),
-          std::cref(*tria),
-          weighting_function);
-      };
+    return [&dof_handler, tria, weighting_function](
+             const typename dealii::Triangulation<dim, spacedim>::cell_iterator
+               &              cell,
+             const CellStatus status) -> unsigned int {
+      return CellWeights<dim, spacedim>::weighting_callback(cell,
+                                                            status,
+                                                            std::cref(
+                                                              dof_handler),
+                                                            std::cref(*tria),
+                                                            weighting_function);
+    };
   }
 
 
@@ -164,7 +162,7 @@ namespace parallel
   unsigned int
   CellWeights<dim, spacedim>::weighting_callback(
     const typename dealii::Triangulation<dim, spacedim>::cell_iterator &cell_,
-    const typename dealii::Triangulation<dim, spacedim>::CellStatus     status,
+    const CellStatus                                                    status,
     const DoFHandler<dim, spacedim> &                 dof_handler,
     const parallel::TriangulationBase<dim, spacedim> &triangulation,
     const typename CellWeights<dim, spacedim>::WeightingFunction
@@ -191,13 +189,13 @@ namespace parallel
     unsigned int fe_index = numbers::invalid_unsigned_int;
     switch (status)
       {
-        case Triangulation<dim, spacedim>::CELL_PERSIST:
-        case Triangulation<dim, spacedim>::CELL_REFINE:
-        case Triangulation<dim, spacedim>::CELL_INVALID:
+        case CELL_PERSIST:
+        case CELL_REFINE:
+        case CELL_INVALID:
           fe_index = cell->future_fe_index();
           break;
 
-        case Triangulation<dim, spacedim>::CELL_COARSEN:
+        case CELL_COARSEN:
 #ifdef DEBUG
           for (const auto &child : cell->child_iterators())
             Assert(child->is_active() && child->coarsen_flag_set(),
