+++ source/distributed/fully_distributed_tria.cc
@@ -445,8 +445,7 @@ namespace parallel
 
       for (const auto &cell : this->active_cell_iterators())
         if (cell->is_locally_owned())
-          this->local_cell_relations.emplace_back(
-            cell, Triangulation<dim, spacedim>::CELL_PERSIST);
+          this->local_cell_relations.emplace_back(cell, CELL_PERSIST);
     }
