+++ tests/mpi/attach_data_02.cc
@@ -37,9 +37,8 @@ template <int dim>
 std::vector<char>
 pack_function(
   const typename parallel::distributed::Triangulation<dim, dim>::cell_iterator
-    &cell,
-  const typename parallel::distributed::Triangulation<dim, dim>::CellStatus
-    status)
+    &              cell,
+  const CellStatus status)
 {
   static unsigned int       some_number = 1;
   std::vector<unsigned int> some_vector(some_number);
@@ -58,17 +57,15 @@ pack_function(
           << " with data size =" << buffer.size() << " accumulated data="
           << std::accumulate(some_vector.begin(), some_vector.end(), 0)
           << " status=";
-  if (status == parallel::distributed::Triangulation<dim, dim>::CELL_PERSIST)
+  if (status == CELL_PERSIST)
     deallog << "PERSIST";
-  else if (status ==
-           parallel::distributed::Triangulation<dim, dim>::CELL_REFINE)
+  else if (status == CELL_REFINE)
     deallog << "REFINE";
-  else if (status ==
-           parallel::distributed::Triangulation<dim, dim>::CELL_COARSEN)
+  else if (status == CELL_COARSEN)
     deallog << "COARSEN";
   deallog << std::endl;
 
-  if (status == parallel::distributed::Triangulation<dim, dim>::CELL_COARSEN)
+  if (status == CELL_COARSEN)
     {
       Assert(cell->has_children(), ExcInternalError());
     }
@@ -85,9 +82,8 @@ template <int dim>
 void
 unpack_function(
   const typename parallel::distributed::Triangulation<dim, dim>::cell_iterator
-    &cell,
-  const typename parallel::distributed::Triangulation<dim, dim>::CellStatus
-                                                                  status,
+    &                                                             cell,
+  const CellStatus                                                status,
   const boost::iterator_range<std::vector<char>::const_iterator> &data_range)
 {
   const unsigned int data_in_bytes =
@@ -111,17 +107,15 @@ unpack_function(
           << " accumulated data="
           << std::accumulate(intdatavector.begin(), intdatavector.end(), 0)
           << " status=";
-  if (status == parallel::distributed::Triangulation<dim, dim>::CELL_PERSIST)
+  if (status == CELL_PERSIST)
     deallog << "PERSIST";
-  else if (status ==
-           parallel::distributed::Triangulation<dim, dim>::CELL_REFINE)
+  else if (status == CELL_REFINE)
     deallog << "REFINE";
-  else if (status ==
-           parallel::distributed::Triangulation<dim, dim>::CELL_COARSEN)
+  else if (status == CELL_COARSEN)
     deallog << "COARSEN";
   deallog << std::endl;
 
-  if (status == parallel::distributed::Triangulation<dim, dim>::CELL_REFINE)
+  if (status == CELL_REFINE)
     {
       Assert(cell->has_children(), ExcInternalError());
     }
