+++ tests/mpi/cell_weights_01_back_and_forth_02.cc
@@ -40,7 +40,7 @@ template <int dim>
 unsigned int
 cell_weight_1(
   const typename parallel::distributed::Triangulation<dim>::cell_iterator &cell,
-  const typename parallel::distributed::Triangulation<dim>::CellStatus status)
+  const CellStatus status)
 {
   return current_cell_weight++;
 }
