+++ tests/mpi/repartition_02.cc
@@ -36,24 +36,21 @@ template <int dim>
 std::vector<char>
 pack_function(
   const typename parallel::distributed::Triangulation<dim, dim>::cell_iterator
-    &cell,
-  const typename parallel::distributed::Triangulation<dim, dim>::CellStatus
-    status)
+    &              cell,
+  const CellStatus status)
 {
   static int some_number = cell->index();
   deallog << "packing cell " << cell->id() << " with data=" << some_number
           << " status=";
-  if (status == parallel::distributed::Triangulation<dim, dim>::CELL_PERSIST)
+  if (status == CELL_PERSIST)
     deallog << "PERSIST";
-  else if (status ==
-           parallel::distributed::Triangulation<dim, dim>::CELL_REFINE)
+  else if (status == CELL_REFINE)
     deallog << "REFINE";
-  else if (status ==
-           parallel::distributed::Triangulation<dim, dim>::CELL_COARSEN)
+  else if (status == CELL_COARSEN)
     deallog << "COARSEN";
   deallog << std::endl;
 
-  if (status == parallel::distributed::Triangulation<dim, dim>::CELL_COARSEN)
+  if (status == CELL_COARSEN)
     {
       Assert(cell->has_children(), ExcInternalError());
     }
@@ -69,9 +66,8 @@ template <int dim>
 void
 unpack_function(
   const typename parallel::distributed::Triangulation<dim, dim>::cell_iterator
-    &cell,
-  const typename parallel::distributed::Triangulation<dim, dim>::CellStatus
-                                                                  status,
+    &                                                             cell,
+  const CellStatus                                                status,
   const boost::iterator_range<std::vector<char>::const_iterator> &data_range)
 {
   const int number = Utilities::unpack<int>(data_range.begin(),
@@ -80,17 +76,15 @@ unpack_function(
 
   deallog << "unpacking cell " << cell->id() << " with data=" << number
           << " status=";
-  if (status == parallel::distributed::Triangulation<dim, dim>::CELL_PERSIST)
+  if (status == CELL_PERSIST)
     deallog << "PERSIST";
-  else if (status ==
-           parallel::distributed::Triangulation<dim, dim>::CELL_REFINE)
+  else if (status == CELL_REFINE)
     deallog << "REFINE";
-  else if (status ==
-           parallel::distributed::Triangulation<dim, dim>::CELL_COARSEN)
+  else if (status == CELL_COARSEN)
     deallog << "COARSEN";
   deallog << std::endl;
 
-  if (status == parallel::distributed::Triangulation<dim, dim>::CELL_REFINE)
+  if (status == CELL_REFINE)
     {
       Assert(cell->has_children(), ExcInternalError());
     }
